main-app/backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
openai
aiosmtplib
passlib[bcrypt]
python-jose[cryptography] 